show how to load the create_tables.cql
