+++ docs-source/docs/modules/microservices-tutorial/examples/shopping-cart-service-scala/ddl-scripts/create_user_tables.cql
@@ -1,0 +1,8 @@
+
+// tag::itemPopularityTables[]
+USE shoppingcart;
+CREATE TABLE IF NOT EXISTS item_popularity (
+  item_id text,
+  count counter,
+  PRIMARY KEY (item_id));
+// end::itemPopularityTables[]
