+++ docs-source/docs/modules/microservices-tutorial/examples/shopping-cart-service-scala/src/main/resources/persistence.conf
@@ -19,12 +19,6 @@ akka {
 
       journal.keyspace = shoppingcart
       snapshot.keyspace = shoppingcart
-
-      # don't use autocreate in production
-      journal.keyspace-autocreate = on
-      journal.tables-autocreate = on
-      snapshot.keyspace-autocreate = on
-      snapshot.tables-autocreate = on
     }
   }
 
@@ -46,4 +40,3 @@ akka.management {
     }
   }
 }
-
